Fixed for merged files creation

+++ fixOut.py
@@ -10,30 +10,31 @@ def main():
         if vol.startswith("Vol "):
             path = base + vol + "/"
             for issue in os.listdir(path):
-                nPath = path + issue + "/"
-                for article in os.listdir(nPath):
-                    if article.startswith("OUT "):
+                if os.path.isdir(path + issue):
+                    nPath = path + issue + "/"
+                    for article in os.listdir(nPath):
+                        if article.startswith("OUT "):
                         
-                        nName = article.replace("OUT ", "")
-                        nName = nName.replace(".pdf", "")
-                        if nName.endswith(" "):
-                            nName = nName.strip()
-                        nName = nName + ".pdf"
-                        print(nName)
-                        start = issue.split(" ")[1]
-                        aNum = nName.split(") - ")[0]
-                        fNum = start + "." + aNum
-                        os.replace(nPath + article, nPath + nName)
-                        location = nPath + nName
-                        f = open(location, 'rb')
-                        pdf = PdfFileReader(f)
-                        info = pdf.getDocumentInfo()
-                        author = info.author
-                        f.close()
-                        fixAllAuthor(fNum, location, author)
-                    else:
-                        if ").pdf" in article:
-                            os.remove(nPath + article)
+                            nName = article.replace("OUT ", "")
+                            nName = nName.replace(".pdf", "")
+                            if nName.endswith(" "):
+                                nName = nName.strip()
+                            nName = nName + ".pdf"
+                            print(nName)
+                            start = issue.split(" ")[1]
+                            aNum = nName.split(") - ")[0]
+                            fNum = start + "." + aNum
+                            os.replace(nPath + article, nPath + nName)
+                            location = nPath + nName
+                            f = open(location, 'rb')
+                            pdf = PdfFileReader(f)
+                            info = pdf.getDocumentInfo()
+                            author = info.author
+                            f.close()
+                            fixAllAuthor(fNum, location, author)
+                        else:
+                            if ").pdf" in article:
+                                os.remove(nPath + article)
                             
 def fixAllAuthor(fNum, location, author):
     path1 = base + "Authors/"
